+++ myVK/myVK/Model/Group/GroupResult.swift
@@ -1,0 +1,10 @@
+// GroupResult.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+import RealmSwift
+
+///  Ответ запроса
+struct GroupResult: Codable {
+    let response: GroupResponse
+}
