+++ myVK/myVK/Model/Photos/Photo.swift
@@ -1,0 +1,16 @@
+// Photo.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+import RealmSwift
+
+/// Данные фото
+final class Photo: Object, Codable {
+    @objc dynamic var ownerID: Int
+    let sizes: [Size]
+
+    enum CodingKeys: String, CodingKey {
+        case ownerID = "owner_id"
+        case sizes
+    }
+}
