+++ myVK/myVK/Model/Photos/PhotoResult.swift
@@ -1,0 +1,10 @@
+// PhotoResult.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+import RealmSwift
+
+///  Ответ запроса
+struct PhotoResult: Codable {
+    let response: PhotoResponse
+}
