+++ myVK/myVK/Model/Photos/PhotoSize.swift
@@ -1,0 +1,10 @@
+// PhotoSize.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+import RealmSwift
+
+///  Ссылка на фото
+final class Size: Object, Codable {
+    @objc dynamic var url: String
+}
