+++ myVK/myVK/Model/User/UserResult.swift
@@ -1,0 +1,10 @@
+// UserResult.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+import RealmSwift
+
+///  Ответ запроса
+struct UserResult: Codable {
+    let response: UserResponse
+}
